+++ features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/quickedit/variant/adapter/factory/ProductVariantPriceAdapterFactoryImpl.kt
@@ -10,13 +10,12 @@ import com.tokopedia.product.manage.common.feature.variant.adapter.model.Product
 import com.tokopedia.product.manage.common.feature.variant.adapter.viewholder.ProductVariantTickerViewHolder
 import com.tokopedia.product.manage.feature.quickedit.variant.adapter.viewholder.ProductVariantPriceViewHolder
 import com.tokopedia.product.manage.feature.quickedit.variant.adapter.viewholder.ProductVariantPriceViewHolder.ProductVariantListener
-import java.math.BigDecimal
 
 class ProductVariantPriceAdapterFactoryImpl(
     private val listener: ProductVariantListener
 ): BaseAdapterTypeFactory(), ProductVariantAdapterFactory {
 
-    private val variantPriceMap: MutableMap<String, BigDecimal> = mutableMapOf()
+    private val variantPriceMap: MutableMap<String, Double> = mutableMapOf()
 
     override fun type(viewModel: ProductVariant): Int = ProductVariantPriceViewHolder.LAYOUT
